CODE RUB: Move 'ExternalAIFileRequest' and 'ExternalAIFileResponse' to ExternalAIFiles folder
