+++ Standard.AI.OpenAI/Brokers/OpenAIs/IOpenAIBroker.AIFiles.cs
@@ -1,9 +1,9 @@
-﻿// ---------------------------------------------------------------------------------- 
+// ---------------------------------------------------------------------------------- 
 // Copyright (c) The Standard Organization, a coalition of the Good-Hearted Engineers 
 // ----------------------------------------------------------------------------------
 
 using System.Threading.Tasks;
-using Standard.AI.OpenAI.Models.Services.Foundations.AIFiles;
+using Standard.AI.OpenAI.Models.Services.Foundations.ExternalAiFiles;
 
 namespace Standard.AI.OpenAI.Brokers.OpenAIs
 {
