+++ Standard.AI.OpenAI/Models/Services/Foundations/ExternalAiFiles/ExternalAIFileResponse.cs
@@ -4,7 +4,7 @@
 
 using Newtonsoft.Json;
 
-namespace Standard.AI.OpenAI.Models.Services.Foundations.AIFiles
+namespace Standard.AI.OpenAI.Models.Services.Foundations.ExternalAiFiles
 {
     internal class ExternalAIFileResponse
     {
